+++ src/routes/(app)/admin/+layout.svelte
@@ -8,15 +8,6 @@
 
 	const i18n = getContext('i18n');
 
-	const SPECIAL_ADMIN_EMAILS = [
-		'ms15138@nyu.edu',
-		'mb484@nyu.edu',
-		'sm11538@nyu.edu',
-		'cg4532@nyu.edu',
-		'jy4421@nyu.edu',
-		'ht2490@nyu.edu'
-	];
-
 	let loaded = false;
 
 	onMount(async () => {
@@ -80,14 +71,13 @@
 								: 'text-gray-600 dark:text-gray-600 hover:text-[#57068c] dark:hover:text-white'} transition"
 							href="/admin/functions">{$i18n.t('Functions')}</a
 						>
-						{#if $user.role === 'admin' && SPECIAL_ADMIN_EMAILS.includes($user.email)}
-							<a
-								class="min-w-fit rounded-full p-1.5 {$page.url.pathname.includes('/admin/settings')
-									? 'text-[#57068c] dark:text-white'
-									: 'text-gray-600 dark:text-gray-600 hover:text-[#57068c] dark:hover:text-white'} transition"
-								href="/admin/settings">{$i18n.t('Settings')}</a
-							>
-						{/if}
+
+						<a
+							class="min-w-fit rounded-full p-1.5 {$page.url.pathname.includes('/admin/settings')
+								? 'text-[#57068c] dark:text-white'
+								: 'text-gray-600 dark:text-gray-600 hover:text-[#57068c] dark:hover:text-white'} transition"
+							href="/admin/settings">{$i18n.t('Settings')}</a
+						>
 					</div>
 				</div>
 			</div>
